backend/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain-text-splitters
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
pytube